remove film param from training
