+++ ivadomed/testing.py
@@ -14,7 +14,8 @@ from ivadomed import inference as imed_inference
 from ivadomed import uncertainty as imed_uncertainty
 from ivadomed.loader import utils as imed_loader_utils
 from ivadomed.object_detection import utils as imed_obj_detect
-from ivadomed.training import get_metadata, store_film_params, save_film_params
+from ivadomed.loader.film import store_film_params, save_film_params
+from ivadomed.training import get_metadata
 from ivadomed.postprocessing import threshold_predictions
 
 cudnn.benchmark = True
